Fix test and rename test path
